+++ PoseA/PoseA/MainView/AnalysisModeRight/Cpp/KalmanFilter3D.hpp
@@ -14,11 +14,11 @@ extern "C" {
 
 typedef struct KalmanFilter3D KalmanFilter3D;
 
-KalmanFilter3D* kalman3d_create(float dt);
-void kalman3d_reset(KalmanFilter3D* filter, float x, float y, float z);
-void kalman3d_update(KalmanFilter3D* filter, float* x, float* y, float* z);
+KalmanFilter3D* kalman3d_create(double dt);
+void kalman3d_reset(KalmanFilter3D* filter, double x, double y, double z);
+void kalman3d_update(KalmanFilter3D* filter, double* x, double* y, double* z);
 void kalman3d_free(KalmanFilter3D* filter);
-void kalman3d_get_state(KalmanFilter3D* filter, float* x, float* y, float* z);
+void kalman3d_get_state(KalmanFilter3D* filter, double* x, double* y, double* z);
 
 #ifdef __cplusplus
 }
